Refactoring: Created module structure for multiple generation plugins
